+++ plugin.video.orange.fr/resources/lib/managers/catchup_manager.py
@@ -1,12 +1,10 @@
 """Catchup TV Manager."""
 
-import xbmc
 import xbmcplugin
 
 from lib.providers import get_provider
 from lib.router import router
-from lib.utils.gui import create_directory_items
-from lib.utils.kodi import build_addon_url
+from lib.utils.gui import create_list_item
 
 
 class CatchupManager:
@@ -19,36 +17,35 @@ class CatchupManager:
     def get_channels(self) -> list:
         """Return channels available for catchup TV."""
         channels = self.provider.get_catchup_channels()
-        directory_items = create_directory_items(channels)
 
-        succeeded = xbmcplugin.addDirectoryItems(router.handle, directory_items, len(directory_items))
-        xbmcplugin.endOfDirectory(router.handle, succeeded)
+        for channel in channels:
+            xbmcplugin.addDirectoryItem(router.handle, channel["path"], create_list_item(channel, True), True)
+
+        xbmcplugin.endOfDirectory(router.handle)
 
     def get_categories(self, catchup_channel_id: str) -> list:
         """Return content categories for the required channel."""
         categories = self.provider.get_catchup_categories(catchup_channel_id)
-        directory_items = create_directory_items(categories)
 
-        succeeded = xbmcplugin.addDirectoryItems(router.handle, directory_items, len(directory_items))
-        xbmcplugin.endOfDirectory(router.handle, succeeded)
+        for category in categories:
+            xbmcplugin.addDirectoryItem(router.handle, category["path"], create_list_item(category, True), True)
+
+        xbmcplugin.endOfDirectory(router.handle)
 
     def get_articles(self, catchup_channel_id: str, category_id: str) -> list:
         """Return content (TV show, movie, etc) for the required channel and category."""
         articles = self.provider.get_catchup_articles(catchup_channel_id, category_id)
-        directory_items = create_directory_items(articles)
 
-        succeeded = xbmcplugin.addDirectoryItems(router.handle, directory_items, len(directory_items))
-        xbmcplugin.endOfDirectory(router.handle, succeeded)
+        for article in articles:
+            xbmcplugin.addDirectoryItem(router.handle, article["path"], create_list_item(article, True), True)
+
+        xbmcplugin.endOfDirectory(router.handle)
 
     def get_videos(self, catchup_channel_id: str, article_id: str) -> list:
         """Return the video list for the required show."""
         videos = self.provider.get_catchup_videos(catchup_channel_id, article_id)
-        directory_items = create_directory_items(videos)
 
-        succeeded = xbmcplugin.addDirectoryItems(router.handle, directory_items, len(directory_items))
-        xbmcplugin.endOfDirectory(router.handle, succeeded)
+        for video in videos:
+            xbmcplugin.addDirectoryItem(router.handle, video["path"], create_list_item(video))
 
-    def play_video(self, video_id: str):
-        """Play catchup video."""
-        player = xbmc.Player()
-        player.play(build_addon_url(f"/catchup-streams/{video_id}"))
+        xbmcplugin.endOfDirectory(router.handle)