+++ plugin.video.orange.fr/resources/lib/managers/stream_manager.py
@@ -1,12 +1,11 @@
 """Video stream manager."""
 
 import inputstreamhelper
-import xbmcgui
 import xbmcplugin
 
 from lib.providers import get_provider
 from lib.router import router
-from lib.utils.gui import create_video_item
+from lib.utils.gui import create_play_item
 from lib.utils.kodi import localize, ok_dialog
 
 
@@ -17,32 +16,29 @@ class StreamManager:
         """Initialize Stream Manager object."""
         self.provider = get_provider()
 
-    def load_live_stream(self, stream_id: str) -> xbmcgui.ListItem:
+    def load_live_stream(self, stream_id: str) -> None:
         """Load live TV stream."""
         stream_info = self.provider.get_live_stream_info(stream_id)
-        if not stream_info:
-            ok_dialog(localize(30900))
-            return
+        self._load_stream(stream_info)
 
-        is_helper = inputstreamhelper.Helper(stream_info["manifest_type"], drm=stream_info["drm"])
-        if not is_helper.check_inputstream():
-            ok_dialog(localize(30901))
-            return
-
-        list_item = create_video_item(stream_info)
-        xbmcplugin.setResolvedUrl(router.handle, True, list_item)
-
-    def load_chatchup_stream(self, stream_id: str) -> xbmcgui.ListItem:
+    def load_chatchup_stream(self, stream_id: str) -> None:
         """Load catchup TV stream."""
         stream_info = self.provider.get_catchup_stream_info(stream_id)
-        if not stream_info:
+        self._load_stream(stream_info)
+
+    def _load_stream(self, stream_info: dict = None) -> None:
+        """Load stream."""
+        if stream_info is None:
             ok_dialog(localize(30900))
+            xbmcplugin.setResolvedUrl(router.handle, False)
             return
 
-        is_helper = inputstreamhelper.Helper(stream_info["manifest_type"], drm=stream_info["drm"])
-        if not is_helper.check_inputstream():
-            ok_dialog(localize(30901))
+        is_helper = inputstreamhelper.Helper(stream_info["manifest_type"], drm=stream_info["license_type"])
+
+        if is_helper.check_inputstream():
+            play_item = create_play_item(stream_info, is_helper.inputstream_addon)
+            xbmcplugin.setResolvedUrl(router.handle, True, play_item)
             return
 
-        list_item = create_video_item(stream_info)
-        xbmcplugin.setResolvedUrl(router.handle, True, list_item)
+        ok_dialog(localize(30901))
+        xbmcplugin.setResolvedUrl(router.handle, False)