+++ plugin.video.orange.fr/resources/lib/utils/gui.py
@@ -3,33 +3,47 @@
 from xbmcgui import ListItem
 
 
-def create_directory_items(data: list) -> list:
-    """Create a list of directory items from data."""
-    items = []
+def create_list_item(item_data: dict, is_folder: bool = False) -> ListItem:
+    """Create a list item from data."""
+    list_item = ListItem(label=item_data.get("label"), path=item_data.get("path"))
 
-    for d in data:
-        list_item = ListItem(label=d["label"], path=d["path"])
+    if "art" in item_data:
+        item_art_data: dict = item_data.get("art", {})
+        list_item.setArt(
+            {
+                "poster": item_art_data.get("poster"),
+                "thumb": item_art_data.get("thumb"),
+            }
+        )
 
-        # list_item.setLabel2("LABEL2")
-        # list_item.setInfo("INFO", {"INFO1": "INFO1", "INFO2": "INFO2"})
+    if not is_folder:
+        list_item.setProperties(
+            {
+                "IsPlayable": "true",
+            }
+        )
 
-        if "art" in d and "thumb" in d["art"]:
-            list_item.setArt({"thumb": d["art"]["thumb"]})
-
-        items.append((d["path"], list_item, bool(d["is_folder"])))
-
-    return items
-
-
-def create_video_item(stream_info: dict) -> ListItem:
-    """Create a video item from stream data."""
-    list_item = ListItem(path=stream_info["path"])
-    list_item.setMimeType(stream_info["mime_type"])
-    list_item.setContentLookup(False)
-    list_item.setProperty("inputstream", "inputstream.adaptive")
-    list_item.setProperty("inputstream.adaptive.play_timeshift_buffer", "true")
-    list_item.setProperty("inputstream.adaptive.manifest_config", '{"timeshift_bufferlimit":14400}')
-    list_item.setProperty("inputstream.adaptive.license_type", stream_info["license_type"])
-    list_item.setProperty("inputstream.adaptive.license_key", stream_info["license_key"])
+    if "info" in item_data:
+        item_info_data: dict = item_data.get("info", {})
+        video_info_tag = list_item.getVideoInfoTag()
+        video_info_tag.setDuration(item_info_data.get("duration"))
+        video_info_tag.setGenres(item_info_data.get("genres"))
+        video_info_tag.setPlot(item_info_data.get("plot"))
+        video_info_tag.setYear(item_info_data.get("year"))
 
     return list_item
+
+
+def create_play_item(stream_info: dict, inputstream_addon: str) -> ListItem:
+    """Create a play item from stream data."""
+    play_item = ListItem(path=stream_info["path"])
+    play_item.setContentLookup(False)
+    play_item.setMimeType(stream_info["mime_type"])
+
+    play_item.setProperty("inputstream", inputstream_addon)
+    # play_item.setProperty("inputstream.adaptive.play_timeshift_buffer", "true")
+    # play_item.setProperty("inputstream.adaptive.manifest_config", '{"timeshift_bufferlimit":14400}')
+    play_item.setProperty("inputstream.adaptive.license_type", stream_info["license_type"])
+    play_item.setProperty("inputstream.adaptive.license_key", stream_info["license_key"])
+
+    return play_item